refactor: split sqlite.go into domain-specific files
Split the monolithic 1558-line sqlite.go into 11 focused files:
- sqlite.go: Store struct, New(), Checkpoint(), Close()
- migrate.go: schema migrations and generateShareToken()
- users.go: UserStore + TierStore methods
- searches.go: SearchStore methods
- dedup.go: DedupStore methods
- notifications.go: NotificationQueue methods
- prices.go: PriceTracker methods
- listings.go: ListingStore + SavedListingStore + HiddenListingStore
- digest.go: DigestStore + DailyDigestStore
- catalog.go: CatalogStore methods
- market.go: MarketStore methods

No logic changes — purely mechanical file reorganization.

Closes #246
